Some fixes in stubs

+++ stubs/TDLib/BaseJsonClient.php
@@ -11,26 +11,15 @@ class BaseJsonClient {
     {
     }
 
-    /**
-     * @param string $query
-     * @return string
-     */
     public function execute(string $query):string
     {
     }
 
-    /**
-     * @param string $query
-     */
     public function send(string $query)
     {
     }
 
-    /**
-     * @param float $timeout
-     * @return string
-     */
-    public function receive(float $timeout):string
+    public function receive(int $timeout):string
     {
     }
 }
